Modified the getHeadCount API and the headcount function used in Home to allow for getting and displaying headcount
